WIP: First draft for Moving EB feature (#54)
This PR is a work in progress.
1. This adds moving eb cases in `Exec/MovingEBCases`. Currently a moving
plane is done with 3 levels of refinement as shown in the gif.
2. It uses a flag - `USE_MOVING_EB` to enclose portions of the codes
which will be enabled only for moving EB cases.
3. The issue of plotting with some undefined values (for certain
plotfiles) exist, and happen **only** for some moving eb cases. But it
is currently fixed with a routine `ZeroingOutForPlotting`, which makes
zero any value less than `1e-12`. This will be fixed in a later PR.
4. Running cases with `fixed_dt` is enabled.

![MovingPlane_Shock_3Level](https://github.com/AMReX-Codes/CAMR/assets/34353555/65ccc6a7-228e-4bc2-bf3f-a595adb9186e)

---------

Co-authored-by: Mahesh Natarajan <mahesh@dogora.dhcp.lbl.gov>